# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders": {
      "post": {
        "description": "Market and limit orders, with partial fills. A GTC limit remainder rests on the book with its funds or shares reserved; an IOC remainder is cancelled and the reservation released.",
        "operationId": "OrdersController_place",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Unique per order — a UUID is ideal. Replaying a key returns the original response instead of placing a second order.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED / IDEMPOTENCY_KEY_REQUIRED"
          },
          "409": {
            "description": "IDEMPOTENT_REQUEST_IN_FLIGHT"
          },
          "422": {
            "description": "INSUFFICIENT_FUNDS / INSUFFICIENT_SHARES / NO_LIQUIDITY"
          },
          "423": {
            "description": "CIRCUIT_BREAKER_TRIPPED / MARKET_HALTED"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Place an order",
        "tags": [
          "Orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [
          {
            "name": "symbol",
            "required": false,
            "in": "query",
            "schema": {
              "example": "vSOL",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "OPEN",
                "PARTIALLY_FILLED",
                "FILLED",
                "CANCELLED",
                "REJECTED"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List your orders, newest first",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/trades": {
      "get": {
        "description": "Every fill, with the position, cost basis and realised P&L that followed it.",
        "operationId": "OrdersController_trades",
        "parameters": [
          {
            "name": "symbol",
            "required": false,
            "in": "query",
            "schema": {
              "example": "vSOL",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TradeDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List your executions",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          },
          "404": {
            "description": "ORDER_NOT_FOUND"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch one order with its fills",
        "tags": [
          "Orders"
        ]
      },
      "delete": {
        "description": "Releases the unfilled reservation. Filled quantity is unaffected.",
        "operationId": "OrdersController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          },
          "404": {
            "description": "ORDER_NOT_FOUND"
          },
          "409": {
            "description": "ORDER_NOT_CANCELLABLE"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a resting order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/portfolio": {
      "get": {
        "description": "Holdings with weighted-average cost basis, mark-to-market value, and realised and unrealised P&L, plus cash and account-level totals.",
        "operationId": "PortfolioController_get",
        "parameters": [
          {
            "name": "verify",
            "required": false,
            "in": "query",
            "description": "Also recompute from the raw ledger and report any drift.",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current portfolio",
        "tags": [
          "Portfolio"
        ]
      }
    },
    "/portfolio/history": {
      "get": {
        "description": "Rebuilds holdings, cost basis and P&L exactly as they stood at `at`, valued at the prices that were actually printed then. Reads the running balance carried on the newest ledger entry and the position snapshot carried on the newest fill at or before that instant, so the cost is three index seeks rather than a replay of the account history. Pass `verify=true` to also recompute the whole thing from raw ledger deltas and report any drift.",
        "operationId": "PortfolioController_history",
        "parameters": [
          {
            "name": "verify",
            "required": false,
            "in": "query",
            "description": "Also recompute from the raw ledger and report any drift.",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "at",
            "required": true,
            "in": "query",
            "description": "ISO-8601 instant to reconstruct the portfolio at.",
            "schema": {
              "example": "2026-08-27T12:00:00.000Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioDto"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED — `at` missing, malformed or in the future"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reconstruct the portfolio at a past instant",
        "tags": [
          "Portfolio"
        ]
      }
    },
    "/portfolio/timeline": {
      "get": {
        "description": "Point-in-time reconstruction sampled across a window. Affordable only because each point is an index seek rather than a history replay.",
        "operationId": "PortfolioController_timeline",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO-8601 start. Defaults to the account’s first activity.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO-8601 end. Defaults to now.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "points",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 2,
              "maximum": 200,
              "default": 24,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquityCurveDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Equity curve",
        "tags": [
          "Portfolio"
        ]
      }
    },
    "/portfolio/ledger": {
      "get": {
        "description": "The raw double-entry ledger behind every balance, newest first.",
        "operationId": "PortfolioController_ledger",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 100,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LedgerEntryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Account statement",
        "tags": [
          "Portfolio"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pre-IPO Trading Platform API",
    "description": "Trade tokenized shares of fictional pre-IPO companies against a stablecoin balance.\n\n**Money formatting** — prices, quantities and cash are returned as decimal *strings*\nto avoid IEEE-754 rounding on the client. Send them the same way.\n\n**Errors** — every failure returns\n`{ \"error\": { \"code\", \"message\", \"details\" }, \"path\", \"timestamp\" }`.\nSwitch on `error.code`, never on the message.\n\n**Idempotency** — `POST /orders` requires an `Idempotency-Key` header. Replaying a key\nreturns the original response instead of placing a second order.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PlaceOrderDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "example": "vSOL"
          },
          "side": {
            "type": "string",
            "enum": [
              "BUY",
              "SELL"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "MARKET",
              "LIMIT"
            ]
          },
          "quantity": {
            "type": "string",
            "example": "12.5",
            "description": "Share quantity. Required for LIMIT orders and for every SELL."
          },
          "usdAmount": {
            "type": "string",
            "example": "5000.00",
            "description": "USD to spend. MARKET BUY only; mutually exclusive with `quantity`."
          },
          "limitPrice": {
            "type": "string",
            "example": "415.00",
            "description": "Required for LIMIT orders."
          },
          "timeInForce": {
            "type": "string",
            "enum": [
              "IOC",
              "GTC"
            ],
            "description": "Defaults to IOC for MARKET orders and GTC for LIMIT orders. A GTC remainder rests on the book; an IOC remainder is cancelled immediately."
          }
        },
        "required": [
          "symbol",
          "side",
          "type"
        ]
      },
      "FillDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "quantity": {
            "type": "string",
            "example": "2.50000000"
          },
          "price": {
            "type": "string",
            "example": "420.250000"
          },
          "notional": {
            "type": "string",
            "example": "1050.625000"
          },
          "fee": {
            "type": "string",
            "example": "1.050625"
          },
          "liquidityRole": {
            "type": "string",
            "enum": [
              "TAKER",
              "MAKER"
            ]
          },
          "counterpartyType": {
            "type": "string",
            "enum": [
              "USER",
              "SYNTHETIC"
            ],
            "description": "USER when matched against another trader, SYNTHETIC against market-maker depth."
          },
          "at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "quantity",
          "price",
          "notional",
          "fee",
          "liquidityRole",
          "counterpartyType",
          "at"
        ]
      },
      "OrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "symbol": {
            "type": "string",
            "example": "vSOL"
          },
          "side": {
            "type": "string",
            "enum": [
              "BUY",
              "SELL"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "MARKET",
              "LIMIT"
            ]
          },
          "timeInForce": {
            "type": "string",
            "enum": [
              "IOC",
              "GTC"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "PARTIALLY_FILLED",
              "FILLED",
              "CANCELLED",
              "REJECTED"
            ],
            "description": "OPEN and PARTIALLY_FILLED still have quantity working on the book. FILLED and CANCELLED are terminal."
          },
          "limitPrice": {
            "type": "object",
            "nullable": true,
            "example": "415.000000"
          },
          "quantity": {
            "type": "string",
            "example": "12.50000000"
          },
          "filledQuantity": {
            "type": "string",
            "example": "8.00000000"
          },
          "remainingQuantity": {
            "type": "string",
            "example": "4.50000000"
          },
          "averageFillPrice": {
            "type": "string",
            "example": "420.250000",
            "description": "Volume-weighted average fill price."
          },
          "filledNotional": {
            "type": "string",
            "example": "3362.000000"
          },
          "feesPaid": {
            "type": "string",
            "example": "3.362000"
          },
          "reservedCash": {
            "type": "string",
            "example": "1875.000000",
            "description": "Cash still earmarked for this order."
          },
          "reservedQuantity": {
            "type": "string",
            "example": "0.00000000",
            "description": "Shares still earmarked for this order."
          },
          "rejectReason": {
            "type": "object",
            "nullable": true
          },
          "fills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FillDto"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "symbol",
          "side",
          "type",
          "timeInForce",
          "status",
          "limitPrice",
          "quantity",
          "filledQuantity",
          "remainingQuantity",
          "averageFillPrice",
          "filledNotional",
          "feesPaid",
          "reservedCash",
          "reservedQuantity",
          "rejectReason",
          "fills",
          "createdAt",
          "updatedAt"
        ]
      },
      "TradeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "quantity": {
            "type": "string",
            "example": "2.50000000"
          },
          "price": {
            "type": "string",
            "example": "420.250000"
          },
          "notional": {
            "type": "string",
            "example": "1050.625000"
          },
          "fee": {
            "type": "string",
            "example": "1.050625"
          },
          "liquidityRole": {
            "type": "string",
            "enum": [
              "TAKER",
              "MAKER"
            ]
          },
          "counterpartyType": {
            "type": "string",
            "enum": [
              "USER",
              "SYNTHETIC"
            ],
            "description": "USER when matched against another trader, SYNTHETIC against market-maker depth."
          },
          "at": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "symbol": {
            "type": "string",
            "example": "vSOL"
          },
          "side": {
            "type": "string",
            "enum": [
              "BUY",
              "SELL"
            ]
          },
          "positionAfter": {
            "type": "string",
            "example": "10.00000000",
            "description": "Position size after this fill."
          },
          "averageCostAfter": {
            "type": "string",
            "example": "418.200000",
            "description": "Average cost after this fill."
          },
          "realizedPnlAfter": {
            "type": "string",
            "example": "125.500000",
            "description": "Cumulative realised P&L after this fill."
          }
        },
        "required": [
          "id",
          "quantity",
          "price",
          "notional",
          "fee",
          "liquidityRole",
          "counterpartyType",
          "at",
          "orderId",
          "symbol",
          "side",
          "positionAfter",
          "averageCostAfter",
          "realizedPnlAfter"
        ]
      },
      "CashDto": {
        "type": "object",
        "properties": {
          "available": {
            "type": "string",
            "example": "223343.240784",
            "description": "Spendable stablecoin."
          },
          "reserved": {
            "type": "string",
            "example": "1251.250000",
            "description": "Earmarked for resting BUY orders."
          },
          "total": {
            "type": "string",
            "example": "224594.490784"
          }
        },
        "required": [
          "available",
          "reserved",
          "total"
        ]
      },
      "HoldingDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "example": "vSOL"
          },
          "name": {
            "type": "string",
            "example": "Solace AI"
          },
          "quantity": {
            "type": "string",
            "example": "45.66966000",
            "description": "Total shares held, reserved included."
          },
          "availableQuantity": {
            "type": "string",
            "example": "40.66966000",
            "description": "Freely sellable shares."
          },
          "reservedQuantity": {
            "type": "string",
            "example": "5.00000000",
            "description": "Shares earmarked for resting SELL orders."
          },
          "averageCost": {
            "type": "string",
            "example": "437.489967",
            "description": "Weighted average cost per share, fees included."
          },
          "costBasis": {
            "type": "string",
            "example": "19980.018067"
          },
          "markPrice": {
            "type": "string",
            "example": "440.120000",
            "description": "Mark price used for valuation."
          },
          "marketValue": {
            "type": "string",
            "example": "20100.140000"
          },
          "unrealizedPnl": {
            "type": "string",
            "example": "120.121933"
          },
          "unrealizedPnlBps": {
            "type": "number",
            "example": 60,
            "description": "Unrealised P&L against cost basis, in basis points."
          },
          "realizedPnl": {
            "type": "string",
            "example": "0.000000",
            "description": "Cumulative realised P&L in this asset."
          },
          "weightBps": {
            "type": "number",
            "example": 812,
            "description": "Share of total portfolio equity, in basis points."
          }
        },
        "required": [
          "symbol",
          "name",
          "quantity",
          "availableQuantity",
          "reservedQuantity",
          "averageCost",
          "costBasis",
          "markPrice",
          "marketValue",
          "unrealizedPnl",
          "unrealizedPnlBps",
          "realizedPnl",
          "weightBps"
        ]
      },
...
      "ReconciliationDto": {
        "type": "object",
        "properties": {
          "consistent": {
            "type": "boolean",
            "description": "True when the O(log n) running-balance read and the independent full ledger fold agree exactly."
          },
          "cashDrift": {
            "type": "string",
            "example": "0.000000",
            "description": "Cash difference between the two methods."
          },
          "positionDrift": {
            "example": [],
            "description": "Symbols where the snapshot and the fill replay disagree. Empty means clean.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ledgerEntriesFolded": {
            "type": "number",
            "example": 3,
            "description": "Ledger entries folded during verification."
          },
          "fillsReplayed": {
            "type": "number",
            "example": 4,
            "description": "Fills replayed during verification."
          }
        },
        "required": [
          "consistent",
          "cashDrift",
          "positionDrift",
          "ledgerEntriesFolded",
          "fillsReplayed"
        ]
      },
      "PortfolioDto": {
        "type": "object",
        "properties": {
          "asOf": {
            "type": "string",
            "example": "2026-08-27T15:30:00.000Z",
            "description": "The instant this portfolio represents."
          },
          "mode": {
            "type": "string",
            "enum": [
              "LIVE",
              "HISTORICAL"
            ],
            "description": "HISTORICAL when reconstructed at a past timestamp."
          },
          "cash": {
            "$ref": "#/components/schemas/CashDto"
          },
          "holdings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HoldingDto"
            }
          },
          "totals": {
            "$ref": "#/components/schemas/PortfolioTotalsDto"
          },
          "reconciliation": {
            "description": "Present when `verify=true`. Proves the fast path matches a full recomputation.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReconciliationDto"
              }
            ]
          }
        },
        "required": [
          "asOf",
          "mode",
          "cash",
          "holdings",
          "totals"
        ]
      },
...
      "EquityCurveDto": {
        "type": "object",
        "properties": {
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EquityPointDto"
            }
          },
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          }
        },
        "required": [
          "points",
          "from",
          "to"
        ]
      },
      "LedgerEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1042"
          },
          "account": {
            "type": "string",
            "enum": [
              "CASH",
              "CASH_RESERVED",
              "POSITION",
              "POSITION_RESERVED"
            ]
          },
          "symbol": {
            "type": "object",
            "nullable": true,
            "example": "vSOL"
          },
          "delta": {
            "type": "string",
            "example": "-19980.018067"
          },
          "balanceAfter": {
            "type": "string",
            "example": "223343.240784"
          },
          "entryType": {
            "type": "string",
            "example": "TRADE_BUY"
          },
          "reference": {
            "type": "object",
            "nullable": true
          },
          "memo": {
            "type": "string",
            "example": "BUY 45.66966000 vSOL @ 437.489967"
          },
          "at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "account",
          "symbol",
          "delta",
          "balanceAfter",
          "entryType",
          "reference",
          "memo",
          "at"
        ]
      },
...
    }
  }
}